requirements.txt
pandas
textdistance
joblib
numpy
scikit-learn
matplotlib
tqdm
paddlepaddle